Separate the node stack from the graph
In the previous version, having multiple nodes depend on the same
could create incorrect graphs, because the new copy of the node
would overwrite the previous version in the graph.

We now separate the stack (which is used to compute dependencies)
and the graph (in which we *update* dependencies and done status,
but nodes with the same name are effectively the same).